refactor: rename ColorCube to Swatch
